# Performance tweaking and improvements