--- a/src/TurnerSoftware.RobotsExclusionTools/Tokenization/TokenMatch.cs
+++ b/src/TurnerSoftware.RobotsExclusionTools/Tokenization/TokenMatch.cs
@@ -9,6 +9,6 @@ namespace TurnerSoftware.RobotsExclusionTools.Tokenization
 		public bool IsMatch { get; set; }
 		public TokenType TokenType { get; set; }
 		public string Value { get; set; }
-		public string RemainingText { get; set; }
+		public int MatchLength { get; set; }
 	}
 }