--- a/src/TurnerSoftware.RobotsExclusionTools/Tokenization/TokenPattern.cs
+++ b/src/TurnerSoftware.RobotsExclusionTools/Tokenization/TokenPattern.cs
@@ -8,14 +8,14 @@ namespace TurnerSoftware.RobotsExclusionTools.Tokenization
 	public class TokenPattern
 	{
 		public TokenType Token { get; }
-		public TokenType[] Preceeding { get; }
-		public TokenType[] Succeeding { get; }
+		public IReadOnlyCollection<TokenType> Preceeding { get; }
+		public IReadOnlyCollection<TokenType> Succeeding { get; }
 
 		public TokenPattern(TokenType token, TokenType[] preceeding, TokenType[] succeeding)
 		{
 			Token = token;
-			Preceeding = preceeding ?? new TokenType[0];
-			Succeeding = succeeding ?? new TokenType[0];
+			Preceeding = preceeding ?? Array.Empty<TokenType>();
+			Succeeding = succeeding ?? Array.Empty<TokenType>();
 		}
 	}
 }