--- a/src/TurnerSoftware.RobotsExclusionTools/Tokenization/Tokenizers/RobotsPageTokenizer.cs
+++ b/src/TurnerSoftware.RobotsExclusionTools/Tokenization/Tokenizers/RobotsPageTokenizer.cs
@@ -6,7 +6,7 @@ namespace TurnerSoftware.RobotsExclusionTools.Tokenization.Tokenizers
 {
 	public class RobotsPageTokenizer : TokenizerBase
 	{
-		private static readonly IEnumerable<TokenDefinition> TokenDefinitions = new List<TokenDefinition>
+		private static readonly IEnumerable<TokenDefinition> TokenDefinitions = new []
 		{
 			//Based on the same character restriction rules as the RobotsFileTokenizer
 			new TokenDefinition(TokenType.Field, @"^[\x21\x23-\x27\x2a\x2b\x2d\x2e\x41-\x5a\x5e-\x7a\x7c\x7e]+(?=:[ ])"),