--- a/src/TurnerSoftware.RobotsExclusionTools/Tokenization/Tokenizers/RobotsPageTokenizer.cs
+++ b/src/TurnerSoftware.RobotsExclusionTools/Tokenization/Tokenizers/RobotsPageTokenizer.cs
@@ -9,11 +9,11 @@ namespace TurnerSoftware.RobotsExclusionTools.Tokenization.Tokenizers
 		private static readonly IEnumerable<TokenDefinition> TokenDefinitions = new []
 		{
 			//Based on the same character restriction rules as the RobotsFileTokenizer
-			new TokenDefinition(TokenType.Field, @"^[\x21\x23-\x27\x2a\x2b\x2d\x2e\x41-\x5a\x5e-\x7a\x7c\x7e]+(?=:[ ])"),
-			new TokenDefinition(TokenType.FieldValueDelimiter, "^:[ ]"),
-			new TokenDefinition(TokenType.Value, @"^[^\x0A\x0D#,]+"),
-			new TokenDefinition(TokenType.NewLine, @"^\x0D?\x0A"),
-			new TokenDefinition(TokenType.ValueDelimiter, "^,[ ]+")
+			new TokenDefinition(TokenType.Field, @"\G[\x21\x23-\x27\x2a\x2b\x2d\x2e\x41-\x5a\x5e-\x7a\x7c\x7e]+(?=:[ ])"),
+			new TokenDefinition(TokenType.FieldValueDelimiter, @"\G:[ ]"),
+			new TokenDefinition(TokenType.Value, @"\G[^\x0A\x0D#,]+"),
+			new TokenDefinition(TokenType.NewLine, @"\G\x0D?\x0A"),
+			new TokenDefinition(TokenType.ValueDelimiter, @"\G,[ ]+")
 		};
 		
 		protected override IEnumerable<TokenDefinition> GetTokenDefinitions()