… tests/TurnerSoftware.RobotsExclusionTools.Benchmarks/StreamingVsStringTokenization.cs