This adds better support for measuring improvements in the benchmark while also improving where possible.

Baseline for updated StreamVsString benchmark:

```
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.18362.900 (1903/May2019Update/19H1)
Intel Core i7-6700HQ CPU 2.60GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.1.301
  [Host]   : .NET Core 3.1.5 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.27001), X64 RyuJIT
  ShortRun : .NET Framework 4.8 (4.8.4180.0), X64 RyuJIT

Job=ShortRun  IterationCount=3  LaunchCount=1
WarmupCount=3
```
|     Method |       Runtime |     Mean |     Error |    StdDev | Ratio | RatioSD |    Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------- |-------------- |---------:|----------:|----------:|------:|--------:|---------:|------:|------:|----------:|
| FromStream |    .NET 4.6.1 | 1.460 ms | 0.2646 ms | 0.0145 ms |  1.00 |    0.00 | 296.8750 |     - |     - | 916.12 KB |
| FromStream | .NET Core 2.1 | 1.330 ms | 0.2976 ms | 0.0163 ms |  0.91 |    0.01 | 265.6250 |     - |     - | 818.09 KB |
| FromStream | .NET Core 3.1 | 1.319 ms | 0.2899 ms | 0.0159 ms |  0.90 |    0.01 | 263.6719 |     - |     - | 809.97 KB |
|            |               |          |           |           |       |         |          |       |       |           |
| FromString |    .NET 4.6.1 | 1.408 ms | 0.1245 ms | 0.0068 ms |  1.00 |    0.00 | 296.8750 |     - |     - | 913.78 KB |
| FromString | .NET Core 2.1 | 1.314 ms | 0.0511 ms | 0.0028 ms |  0.93 |    0.00 | 265.6250 |     - |     - |  818.2 KB |
| FromString | .NET Core 3.1 | 1.304 ms | 0.3364 ms | 0.0184 ms |  0.93 |    0.02 | 263.6719 |     - |     - | 810.63 KB |