@Turnerj 2020-06-13
Results:
- FromStream ~15% faster with ~20% less allocations
- FromString ~18% faster with ~20% less allocations

Benchmark Results
```
|     Method |       Runtime |     Mean |     Error |    StdDev | Ratio | RatioSD |    Gen 0 |  Gen 1 | Gen 2 | Allocated |
|----------- |-------------- |---------:|----------:|----------:|------:|--------:|---------:|-------:|------:|----------:|
| FromStream |    .NET 4.6.1 | 1.244 ms | 0.1416 ms | 0.0078 ms |  1.00 |    0.00 | 230.4688 |      - |     - | 708.33 KB |
| FromStream | .NET Core 2.1 | 1.083 ms | 0.0592 ms | 0.0032 ms |  0.87 |    0.01 | 218.7500 | 1.9531 |     - | 675.27 KB |
| FromStream | .NET Core 3.1 | 1.126 ms | 0.3056 ms | 0.0168 ms |  0.90 |    0.02 | 218.7500 | 1.9531 |     - |  670.4 KB |
|            |               |          |           |           |       |         |          |        |       |           |
| FromString |    .NET 4.6.1 | 1.161 ms | 0.0749 ms | 0.0041 ms |  1.00 |    0.00 | 228.5156 |      - |     - | 706.01 KB |
| FromString | .NET Core 2.1 | 1.064 ms | 0.2479 ms | 0.0136 ms |  0.92 |    0.01 | 218.7500 | 1.9531 |     - | 675.38 KB |
| FromString | .NET Core 3.1 | 1.089 ms | 0.1227 ms | 0.0067 ms |  0.94 |    0.01 | 218.7500 | 1.9531 |     - | 671.06 KB |
```